Merge from original

+++ lib/classes/record/adapter.php
@@ -19,6 +19,7 @@ use Alchemy\Phrasea\Core\Event\Record\OriginalNameChangedEvent;
 use Alchemy\Phrasea\Core\Event\Record\RecordEvent;
 use Alchemy\Phrasea\Core\Event\Record\RecordEvents;
 use Alchemy\Phrasea\Core\Event\Record\StatusChangedEvent;
+use Alchemy\Phrasea\Core\Event\Record\SubdefinitionCreateEvent;
 use Alchemy\Phrasea\Core\PhraseaTokens;
 use Alchemy\Phrasea\Databox\Subdef\MediaSubdefRepository;
 use Alchemy\Phrasea\Filesystem\FilesystemService;
@@ -31,6 +32,7 @@ use Alchemy\Phrasea\Model\Entities\User;
 use Alchemy\Phrasea\Model\RecordInterface;
 use Alchemy\Phrasea\Model\Serializer\CaptionSerializer;
 use Alchemy\Phrasea\Record\RecordReference;
+use Alchemy\Phrasea\SearchEngine\Elastic\Indexer\Record\Hydrator\GpsPosition;
 use Alchemy\Phrasea\SearchEngine\SearchEngineInterface;
 use Alchemy\Phrasea\SearchEngine\SearchEngineOptions;
 use Doctrine\DBAL\Connection;
@@ -283,7 +285,7 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         $this->getDataboxConnection()->executeUpdate($sql, ['type' => $type, 'record_id' => $this->getRecordId()]);
 
         if ($old_type !== $type) {
-            $this->rebuild_subdefs();
+            $this->dispatch(RecordEvents::SUBDEFINITION_CREATE, new SubdefinitionCreateEvent($this));
         }
 
         $this->type = $type;
@@ -340,7 +342,8 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
             array(':mime' => $mime, ':record_id' => $this->getRecordId())
         )) {
 
-            $this->rebuild_subdefs();
+            $this->dispatch(RecordEvents::SUBDEFINITION_CREATE, new SubdefinitionCreateEvent($this));
+
             $this->delete_data_from_cache();
         }
 
@@ -519,19 +522,23 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
     /**
      *
      * @param  collection     $collection
-     * @param  appbox         $appbox
+     * @param  appbox         $appbox       WTF this parm is useless
      * @return record_adapter
+     *
      */
-    public function move_to_collection(collection $collection, appbox $appbox)
+    public function move_to_collection(collection $collection, appbox $appbox = null)
     {
         if ($this->getCollection()->get_base_id() === $collection->get_base_id()) {
             return $this;
         }
 
+        $coll_id_from = $this->getCollectionId();
+        $coll_id_to = $collection->get_coll_id();
+
         $sql = "UPDATE record SET moddate = NOW(), coll_id = :coll_id WHERE record_id =:record_id";
 
         $params = [
-            ':coll_id'   => $collection->get_coll_id(),
+            ':coll_id'   => $coll_id_to,
             ':record_id' => $this->getRecordId(),
         ];
 
@@ -540,11 +547,12 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         $stmt->closeCursor();
 
         $this->base_id = $collection->get_base_id();
-
-        $this->app['phraseanet.logger']($this->getDatabox())
-            ->log($this, Session_Logger::EVENT_MOVE, $collection->get_coll_id(), '');
+        $this->collection_id = $coll_id_to;
 
         $this->delete_data_from_cache();
+
+        $this->app['phraseanet.logger']($this->getDatabox())
+            ->log($this, Session_Logger::EVENT_MOVE, $collection->get_coll_id(), '', $coll_id_from);
 
         $this->dispatch(RecordEvents::COLLECTION_CHANGED, new CollectionChangedEvent($this));
 
@@ -747,6 +755,38 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         }
 
         return $this->technical_data;
+    }
+
+    /**
+     * Return an array containing GPSPosition
+     *
+     * @return array
+     */
+    public function getPositionFromTechnicalInfos()
+    {
+        $positionTechnicalField = [
+            media_subdef::TC_DATA_LATITUDE,
+            media_subdef::TC_DATA_LONGITUDE
+        ];
+        $position = [];
+
+        foreach($positionTechnicalField as $field){
+            $fieldData = $this->get_technical_infos($field);
+
+            if($fieldData){
+                $position[$field] = $fieldData->getValue();
+            }
+        }
+
+        if(count($position) == 2){
+            return [
+                'isCoordComplete' => 1,
+                'latitude' => $position[media_subdef::TC_DATA_LATITUDE],
+                'longitude' => $position[media_subdef::TC_DATA_LONGITUDE]
+            ];
+        }
+
+        return ['isCoordComplete' => 0, 'latitude' => 'false', 'longitude' => 'false'];
     }
 
     /**
@@ -1015,10 +1055,13 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
             }
         }
 
-        if (trim($params['meta_id']) !== '') {
-            $tmp_val = trim($params['value']);
+        $tmp_val = trim($params['value']);
 
-            $caption_field_value = $caption_field->get_value($params['meta_id']);
+        if (trim($params['meta_id']) !== '') {
+
+            if(is_null($caption_field_value = $caption_field->get_value($params['meta_id']))) {
+                return $this;
+            }
 
             if ($tmp_val === '') {
                 $caption_field_value->delete();
@@ -1029,8 +1072,11 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
                     $caption_field_value->setVocab($vocab, $vocab_id);
                 }
             }
-        } else {
-            $caption_field_value = caption_Field_Value::create($this->app, $databox_field, $this, $params['value'], $vocab, $vocab_id);
+        }
+        else {
+            if($tmp_val !== '') {
+                caption_Field_Value::create($this->app, $databox_field, $this, $params['value'], $vocab, $vocab_id);
+            }
         }
 
         return $this;
@@ -1068,6 +1114,7 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         $this->set_xml($xml);
         unset($xml);
 
+        $this->write_metas();
         $this->dispatch(RecordEvents::METADATA_CHANGED, new MetadataChangedEvent($this));
 
         return $this;
@@ -1174,13 +1221,14 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         try {
             $log_id = $app['phraseanet.logger']($collection->get_databox())->get_id();
 
-            $sql = 'INSERT INTO log_docs (id, log_id, date, record_id, action, final, comment)'
-                . ' VALUES (null, :log_id, now(), :record_id, "add", :coll_id,"")';
+            $sql = 'INSERT INTO log_docs (id, log_id, date, record_id, coll_id, action, final, comment)'
+                . ' VALUES (null, :log_id, now(), :record_id, :coll_id, "add", :final, "")';
             $stmt = $connection->prepare($sql);
             $stmt->execute([
                 ':log_id'    => $log_id,
                 ':record_id' => $story_id,
                 ':coll_id'   => $collection->get_coll_id(),
+                ':final'     => $collection->get_coll_id(),
             ]);
             $stmt->closeCursor();
         } catch (\Exception $e) {
@@ -1206,7 +1254,8 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
             . " (coll_id, record_id, parent_record_id, moddate, credate, type, sha256, uuid, originalname, mime)"
             . " VALUES (:coll_id, null, :parent_record_id, NOW(), NOW(), :type, :sha256, :uuid, :originalname, :mime)";
 
-        $stmt = $databox->get_connection()->prepare($sql);
+        $connection = $databox->get_connection();
+        $stmt = $connection->prepare($sql);
 
         $stmt->execute([
             ':coll_id'          => $file->getCollection()->get_coll_id(),
@@ -1219,21 +1268,22 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         ]);
         $stmt->closeCursor();
 
-        $record_id = $databox->get_connection()->lastInsertId();
+        $record_id = $connection->lastInsertId();
 
         $record = new self($app, $databox->get_sbas_id(), $record_id);
 
         try {
             $log_id = $app['phraseanet.logger']($databox)->get_id();
 
-            $sql = "INSERT INTO log_docs (id, log_id, date, record_id, action, final, comment)"
-                . " VALUES (null, :log_id, now(), :record_id, 'add', :coll_id, '')";
+            $sql = "INSERT INTO log_docs (id, log_id, date, record_id, coll_id, action, final, comment)"
+                . " VALUES (null, :log_id, now(), :record_id, :coll_id, 'add', :final, '')";
 
             $stmt = $databox->get_connection()->prepare($sql);
             $stmt->execute([
                 ':log_id'    => $log_id,
                 ':record_id' => $record_id,
                 ':coll_id'   => $file->getCollection()->get_coll_id(),
+                ':final'     => $file->getCollection()->get_coll_id(),
             ]);
             $stmt->closeCursor();
         } catch (\Exception $e) {
@@ -1303,6 +1353,37 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         $this->delete_data_from_cache(self::CACHE_TECHNICAL_DATA);
 
         return $this;
+    }
+
+    /**
+     * Insert or update technical data
+     * $technicalDatas an array of name => value
+     *
+     * @param array $technicalDatas
+     */
+    public function insertOrUpdateTechnicalDatas($technicalDatas)
+    {
+        $technicalFields = media_subdef::getTechnicalFieldsList();
+        $sqlValues = null;
+
+        foreach($technicalDatas as $name => $value){
+            if(array_key_exists($name, $technicalFields)){
+                if(is_null($value)){
+                    $value = 0;
+                }
+                $sqlValues[] = [$this->getRecordId(), $name, $value, $value];
+            }
+        }
+
+        if($sqlValues){
+            $connection = $this->getDataboxConnection();
+            $connection->transactional(function (Connection $connection) use ($sqlValues) {
+                $statement = $connection->prepare('INSERT INTO technical_datas (record_id, name, value) VALUES (?, ?, ?) ON DUPLICATE KEY UPDATE value = ?');
+                array_walk($sqlValues, [$statement, 'execute']);
+            });
+
+            $this->delete_data_from_cache(self::CACHE_TECHNICAL_DATA);
+        }
     }
 
     /**
@@ -1534,14 +1615,15 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
 
     public function log_view($log_id, $referrer, $gv_sit)
     {
-        $sql = "INSERT INTO log_view (id, log_id, date, record_id, referrer, site_id)"
-            . " VALUES (null, :log_id, now(), :rec, :referrer, :site)";
+        $sql = "INSERT INTO log_view (id, log_id, date, record_id, referrer, site_id, coll_id)"
+            . " VALUES (null, :log_id, now(), :rec, :referrer, :site, :collid)";
 
         $params = [
             ':log_id'   => $log_id
             , ':rec'      => $this->getRecordId()
             , ':referrer' => $referrer
-            , ':site'     => $gv_sit,
+            , ':site'     => $gv_sit
+            , ':collid'   => $this->getCollectionId()
         ];
         $stmt = $this->getDataboxConnection()->prepare($sql);
         $stmt->execute($params);
@@ -1600,6 +1682,43 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
         return $records;
     }
 
+    public static function getRecordsByOriginalnameWithExcludedCollIds(databox $databox, $original_name, $caseSensitive = false, $offset_start = 0, $how_many = 10, $excludedCollIds = [])
+    {
+        $offset_start = max(0, (int)$offset_start);
+        $how_many = max(1, (int)$how_many);
+        $collate = $caseSensitive ? 'utf8_bin' : 'utf8_unicode_ci';
+
+        $qb = $databox->get_connection()->createQueryBuilder()
+            ->select('record_id')
+            ->from('record')
+            ->where('originalname = :original_name COLLATE :collate')
+            ;
+
+        $params = ['original_name' => $original_name, 'collate' => $collate];
+        $types  = [];
+
+        if (!empty($excludedCollIds)) {
+            $qb->andWhere($qb->expr()->notIn('coll_id', ':coll_id'));
+
+            $params['coll_id'] = $excludedCollIds;
+            $types[':coll_id'] = Connection::PARAM_INT_ARRAY;
+        }
+
+        $sql = $qb->setFirstResult($offset_start)
+            ->setMaxResults($how_many)
+            ->getSQL()
+            ;
+
+        $rs = $databox->get_connection()->fetchAll($sql, $params, $types);
+
+        $records = [];
+        foreach ($rs as $row) {
+            $records[] = $databox->get_record($row['record_id']);
+        }
+
+        return $records;
+    }
+
     /**
      * @return set_selection|record_adapter[]
      * @throws Exception
@@ -1612,17 +1731,22 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
     }
 
     /**
+     * @param int $offset
+     * @param null|int $max_items
+     *
      * @return set_selection|record_adapter[]
      * @throws Exception
      * @throws \Doctrine\DBAL\DBALException
      */
-    public function getChildren()
+    public function getChildren($offset = 1, $max_items = null)
     {
         if (!$this->isStory()) {
             throw new Exception('This record is not a grouping');
         }
 
-        $selections = $this->getDatabox()->getRecordRepository()->findChildren([$this->getRecordId()]);
+        $user = $this->getAuthenticatedUser();
+
+        $selections = $this->getDatabox()->getRecordRepository()->findChildren([$this->getRecordId()], $user, $offset, $max_items);
 
         return reset($selections);
     }
@@ -1632,7 +1756,9 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
      */
     public function get_grouping_parents()
     {
-        $selections = $this->getDatabox()->getRecordRepository()->findParents([$this->getRecordId()]);
+        $user = $this->getAuthenticatedUser();
+
+        $selections = $this->getDatabox()->getRecordRepository()->findParents([$this->getRecordId()], $user);
 
         return reset($selections);
     }
@@ -1834,5 +1960,16 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
     private function getMediaSubdefRepository()
     {
         return $this->app['provider.repo.media_subdef']->getRepositoryForDatabox($this->getDataboxId());
+    }
+
+    /**
+     * @return User|null
+     */
+    protected function getAuthenticatedUser()
+    {
+        /** @var \Alchemy\Phrasea\Authentication\Authenticator $authenticator */
+        $authenticator = $this->app['authentication'];
+
+        return $authenticator->getUser();
     }
 }
